enhancement

+++ app/src/main/java/com/example/myapplication/models/repo/NewsRepo.kt
@@ -1,67 +1,42 @@
 package com.example.myapplication.models.repo
 
 import android.content.Context
-import android.util.Log
-import android.widget.Toast
-import androidx.room.Room
-import com.example.myapplication.models.POJO.Article
-import com.example.myapplication.models.POJO.NewsResponse
+import com.example.myapplication.models.pojo.Article
+import com.example.myapplication.models.pojo.NewsResponse
 import com.example.myapplication.models.database.AppDatabase
 import com.example.myapplication.models.database.DatabaseGenerator
 import com.example.myapplication.models.service.APIService
 import com.example.myapplication.models.service.ServiceGenerator
-import com.example.myapplication.view.interfaces.Constants
+import com.example.myapplication.Constants
+import com.example.myapplication.Utils
 import com.example.myapplication.view.interfaces.NewsInterface
-import retrofit2.Call
 import retrofit2.Callback
-import retrofit2.Response
 
-class NewsRepo : NewsInterface.NewsModel {
+class NewsRepo(context: Context) : NewsInterface.NewsModel {
 
 
-    private var application: APIService? = null
-    private var appDatabase: AppDatabase? = DatabaseGenerator.INSTANCE
+    private var ApiService: APIService? = null
+    private var appDatabase: AppDatabase? = DatabaseGenerator().getInstance(context)
+
 
     init {
-        application = ServiceGenerator.buildService(APIService::class.java)
+        ApiService = ServiceGenerator.buildService(APIService::class.java)
     }
 
-    override fun getNews(presenter: NewsInterface.NewsPresenter) {
-        presenter.loading()
-        application?.getLatestNews("techcrunch", Constants.API_KEY)
-            ?.enqueue(object : Callback<NewsResponse> {
-                override fun onResponse(
-                    call: Call<NewsResponse>,
-                    response: Response<NewsResponse>
-                ) {
-                    if (response.body()?.status.equals("ok")) {
-                        presenter.doneLoading()
-                        response.body()?.let { presenter.uiAutoUpdate(it.articles) }
-                        response.body()?.let { saveInDB(it.articles) }
-                    } else {
-                        presenter.showError()
-                    }
-                }
-
-                override fun onFailure(call: Call<NewsResponse>, t: Throwable) {
-                    presenter.doneLoading()
-                    presenter.showError()
-                }
-            })
+    override fun getNews(call: Callback<NewsResponse>) {
+        ApiService?.getLatestNews(Constants.API.SOURCE, Constants.API.API_KEY)
+            ?.enqueue(call)
     }
 
-    override fun makeCash(presenter: NewsInterface.NewsPresenter) {
-        val articles = appDatabase?.getArticlesDao()?.getAllArticles()
-        if (articles != null) {
-            presenter.uiAutoUpdate(articles)
-        } else {
-            presenter.cashError()
+    override fun getCache(): List<Article>? {
+        return appDatabase?.getArticlesDao()?.getAllArticles()
+    }
 
+    override fun saveInDB(articles: List<Article>) {
+        if (appDatabase?.getArticlesDao()?.getAllArticles()?.isNotEmpty() == true) {
+            appDatabase?.getArticlesDao()?.deleteAll()
         }
-    }
-
-    fun saveInDB(articles: List<Article>) {
-        appDatabase?.getArticlesDao()?.deleteAll()
+        Utils.log(appDatabase.toString() + " saved in data")
         appDatabase?.getArticlesDao()?.insertArticles(articles)
     }
